Respond to review comments. Add SleepService
